requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
fastapi==0.118.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
